# src/Dle.Control/openapi/dle-control.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Deep Link Engine — control plane",
    "description": "Link, domain, application, tenant and credential management, plus attribution and analytics reporting. Errors are RFC 9457 problem documents whose type member is one of the stable identifiers documented at https://docs.dle.dev/problems/. Writes accept an Idempotency-Key header: the same key with the same body replays the stored response, and the same key with a different body is a conflict.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/tenants/me": {
      "get": {
        "tags": [
          "Tenants"
        ],
        "summary": "Reads the caller's own tenant.",
        "description": "Separate from the identifier route on purpose. Reading one's own tenant is an ordinary,\r\nviewer-level operation; reading somebody else's is instance administration, and conflating\r\nthem would mean every viewer credential carried a route that could address another tenant.",
        "operationId": "GetOwnTenant",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants": {
      "get": {
        "tags": [
          "Tenants"
        ],
        "summary": "Lists every tenant on the instance.",
        "operationId": "ListTenants",
        "parameters": [
          {
            "name": "includeDeleted",
            "in": "query",
            "description": "Whether deactivated tenants are listed too.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfTenantResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tenants"
        ],
        "summary": "Provisions a tenant (FR-241).",
        "description": "The default consent mode is `aggregate_only`, the mode that needs no consent banner. A\r\nfresh tenant therefore never collects more than the legitimate interest basis supports until\r\nsomebody deliberately widens it (§E.6.2, FR-248).",
        "operationId": "CreateTenant",
        "requestBody": {
          "description": "Cancellation token.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{id}": {
      "get": {
        "tags": [
          "Tenants"
        ],
        "summary": "Reads one tenant.",
        "operationId": "GetTenant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The tenant identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Tenants"
        ],
        "summary": "Changes a tenant.",
        "description": "A consent mode change is recorded with the old and the new value. Tightening takes effect on\r\nthe next resolve; widening does not retroactively make already collected data lawful, and the\r\naudit entry is what lets that distinction be reconstructed later (§E.6.2).",
        "operationId": "UpdateTenant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The tenant identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Cancellation token.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tenants"
        ],
        "summary": "Deactivates a tenant.",
        "description": "The row stays. A hard delete would cascade through links, keys and the audit trail, and the\r\naudit trail is the one thing that has to outlive the objects it describes; the soft-delete\r\nfilter is what stops the tenant appearing anywhere it should not.",
        "operationId": "DeleteTenant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The tenant identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/exports/tenant": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Handles an export.",
        "description": "GDPR article 20 portability and the DORA article 30 exit plan are the same feature, and this is it: newline delimited JSON per table in a ZIP, streamed straight out of the database. Credential material — key hashes, webhook secrets, account hashes — is never included, and the manifest says so explicitly rather than leaving the gap to be discovered (FR-249, §E.6.4).",
        "operationId": "ExportTenantData",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTenantRequest": {
        "required": [
          "slug",
          "name"
        ],
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "Short unique identifier used in URLs and logs."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          },
          "consent_mode": {
            "type": [
              "null",
              "string"
            ],
            "description": "Consent mode: `off`, `aggregate_only` or `full`. Defaults to\r\n            `aggregate_only`, the mode that needs no consent banner (§E.6.2, FR-248)."
          }
        },
        "description": "Body of `POST /api/v1/tenants` (FR-241)."
      },
...
      "PagedResponseOfTenantResponse": {
        "required": [
          "items"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TenantResponse"
            },
            "description": "The items on this page."
          },
          "next_cursor": {
            "type": [
              "null",
              "string"
            ],
            "description": "Opaque cursor for the next page, or `null` at the end."
          },
          "total": {
            "type": [
              "null",
              "integer"
            ],
            "description": "Total matching items when the store can supply it cheaply.",
            "format": "int64"
          }
        },
        "description": "A page of results plus the cursor needed to fetch the next one."
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "TenantResponse": {
        "required": [
          "id",
          "slug",
          "name",
          "status",
          "consent_mode",
          "created_at"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Tenant identifier.",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "description": "Short unique identifier."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          },
          "status": {
            "type": "string",
            "description": "Lifecycle status: `active`, `suspended` or `deleted`."
          },
          "consent_mode": {
            "type": "string",
            "description": "Effective consent mode."
          },
          "created_at": {
            "type": "string",
            "description": "Creation instant.",
            "format": "date-time"
          }
        },
        "description": "Representation of a tenant."
      },
...
      "UpdateTenantRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ],
            "description": "New display name."
          },
          "consent_mode": {
            "type": [
              "null",
              "string"
            ],
            "description": "New consent mode: `off`, `aggregate_only` or `full` (FR-248, §E.6.2)."
          },
          "status": {
            "type": [
              "null",
              "string"
            ],
            "description": "New lifecycle status: `active` or `suspended`."
          }
        },
        "description": "Body of `PATCH /api/v1/tenants/{id}`. Every member is optional; only the supplied ones\r\nchange."
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "description": "A control-plane API key, presented as a bearer token or in the X-Api-Key header. The key looks like dle_<prefix>_<secret>; only its Argon2id hash is stored, so a lost key is replaced rather than recovered.",
        "scheme": "bearer"
      },
      "SdkKey": {
        "type": "apiKey",
        "description": "A key embedded in a customer application. It authenticates the two SDK ingestion routes and nothing else, and must be assumed extractable from the application binary.",
        "name": "X-Dle-Sdk-Key",
        "in": "header"
      }
    }
  },
...
}